Move error state to store
