+++ src/Chat/Chat.tsx
@@ -1,6 +1,6 @@
 import "./Chat.scss";
 import React, { useRef, useState } from "react";
-import Card from "../Card";
+import Card from "../components/Card";
 import Timer from "./Timer";
 import type { ChatPayload } from "../types/chatPayload.type";
