+++ src/Connect.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from "react";
-import Card from "./Card";
+import Card from "./components/Card";
 import ConnectionManager from "./utils/connectionManager";
 import type { AnyPayload } from "./types/anyPayload.type";
