+++ src/PlayersList/PlayersList.tsx
@@ -1,4 +1,4 @@
-import Card from "../Card";
+import Card from "../components/Card";
 import { useLobbyStore } from "../utils/store";
 
 const PlayersList = function PlayersList(props: {
